UserScrape/requirements.txt
selenium
azure-storage-blob